+++ src/ray/common/task/task_util.h
@@ -4,7 +4,6 @@
 #include "ray/common/buffer.h"
 #include "ray/common/ray_object.h"
 #include "ray/common/task/task_spec.h"
-#include "ray/core_worker/common.h"
 #include "ray/protobuf/common.pb.h"
 
 namespace ray {
