+++ src/ray/gcs/subscription_executor_test.cc
@@ -45,7 +45,7 @@ class SubscriptionExecutorTest : public AccessorTestBase<ActorID, ActorTableData
     for (size_t i = 0; i < 2; ++i) {
       std::shared_ptr<ActorTableData> actor = std::make_shared<ActorTableData>();
       actor->set_max_reconstructions(1);
-      actor->set_num_lifetimes(0);
+      actor->set_num_restarts(0);
       JobID job_id = JobID::FromInt(i);
       actor->set_job_id(job_id.Binary());
       actor->set_state(ActorTableData::ALIVE);
